v2.018.2
Reverted to last known good commit.
For the love of God, double and triple check that you are always pulling when working, and pushing when you get done even if you're not finished. Merge conflicts were introduced on commit 2.019 that broke all CSProj files and as a result our ability to program wasn't set up right. Be careful in future.